test: split the test file
